Calculate time span for AzureConsumptionPeriod
This commits ensures the correct amount of days data is collected over is written in the generated reports.
Before this commit, when using the DoAzureConsumptionPreviousMonth the AzureConsumptionPeriod param was no longer technically relevant, but was used in reports, e.g. it would read as the last X amount of days was collected based on the param value provided, while the actual amount of days depends on the length of the previous month. (between 28 - 31 days).

With this commit the correct amount of days is used throughout the reports, which lead to some other phrasing changes (e.g. it is not necessarily the 'last' X days the report is collected over).